requirement.txt
asgiref==3.5.1
Django==3.2
django-leaflet==0.28.2
djangorestframework==3.13.1
djangorestframework-gis==0.18
psycopg2==2.9.3
pytz==2022.1
sqlparse==0.4.2
